// frontend/src/types.ts
export interface OptionType {
  readonly value: string;
  readonly label: string;
}